requirements.txt
python-pptx==0.6.23
reportlab==4.0.9
Pillow==10.2.0
pypdf>=4.0
